Update 1.0.6
Fix GetDefinitionIndex on windows
PTaH_GetEconItemViewFromWeapon now works without gamedata
Add PTaH_SpawnItemFromDefIndex - allows you to create objects by their DefinitionIndex with a valid CEconItemView

+++ classes.h
@@ -40,7 +40,7 @@ enum EStickerAttributeType
 class CEconItemDefinition
 {
 public:
-	int GetDefinitionIndex();
+	uint16_t GetDefinitionIndex();
 	int GetLoadoutSlot(int def);
 	int GetNumSupportedStickerSlots();
 };
